refactor image-url-helper
