+++ development/image-url-helper/pkg/promote/promote.go
@@ -9,7 +9,7 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/kyma-project/test-infra/development/image-url-helper/pkg/list"
+	"github.com/kyma-project/test-infra/development/image-url-helper/pkg/common"
 	"gopkg.in/yaml.v3"
 )
 
@@ -22,7 +22,7 @@ type ExcludesList struct {
 	Excludes []string `yaml:"excludes"`
 }
 
-func GetWalkFunc(ResourcesDirectoryClean, targetContainerRegistry, targetTag string, dryRun bool, images, testImages list.ImageMap, excludes ExcludesMap) filepath.WalkFunc {
+func GetWalkFunc(ResourcesDirectoryClean, targetContainerRegistry, targetTag string, dryRun bool, images, testImages common.ImageMap, excludes ExcludesMap) filepath.WalkFunc {
 	return func(path string, info os.FileInfo, err error) error {
 		//pass the error further, this shouldn't ever happen
 		if err != nil {
@@ -45,7 +45,7 @@ func GetWalkFunc(ResourcesDirectoryClean, targetContainerRegistry, targetTag str
 		}
 
 		var parsedFile yaml.Node
-		var parsedImagesFile list.ValueFile
+		var parsedImagesFile common.ValueFile
 		lines := make([]string, 0)
 
 		yamlFile, err := os.Open(path)
@@ -80,7 +80,7 @@ func GetWalkFunc(ResourcesDirectoryClean, targetContainerRegistry, targetTag str
 		}
 
 		// generate list of used images and apprend it to the global list containing images from all values.yaml files
-		list.AppendImagesToMap(parsedImagesFile, images, testImages, "", make(list.ImageToComponents))
+		common.AppendImagesToMap(parsedImagesFile, images, testImages, "", make(common.ImageToComponents))
 
 		globalNode := getYamlNode(parsedFile.Content[0], "global")
 		if globalNode == nil {
